requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
discord==2.3.2
discord.py==2.3.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
Flask==3.0.2
frozenlist==1.4.1
greenlet==1.1.2
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tomli==2.0.1
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
Werkzeug==3.0.1
yarl==1.9.4
zipp==1.0.0
